Move UnstructuredTracePublisher to Diagnostics
