+++ test/unittests/Microsoft.ServiceFabric.Actors.Tests/Diagnostics/ActorFrameworkEventSourceTest.cs
@@ -7,6 +7,7 @@ using Microsoft.ServiceFabric.Services;
 using Moq;
 using Xunit;
 using Xunit.Abstractions;
+using Microsoft.ServiceFabric.Diagnostics.Tracing;
 
 namespace Microsoft.ServiceFabric.Actors.Diagnostics
 {
