+++ test/unittests/Microsoft.ServiceFabric.Diagnostics.Tests/ServiceFabricEventSourceTest.cs
@@ -13,7 +13,6 @@ namespace Microsoft.ServiceFabric.Diagnostics.Tests
 {    
     public abstract class ServiceFabricEventSourceTest
     {
-#if DotNetCoreClr
         public sealed class Class : ServiceFabricEventSourceTest
         {
             [Fact]
@@ -49,55 +48,6 @@ namespace Microsoft.ServiceFabric.Diagnostics.Tests
                 // Restore Writer singleton
                 typeof(TestEventSource).Property<TestEventSource>().Set(new TestEventSource());
             }
-
-            [Theory]
-            [InlineData(1, true, 1)]
-            [InlineData(2, true, 2)]
-            [InlineData(3, true, -1)]
-            public void Constructor_OnLinux_GeneratesEventDescriptorsCorrectly(int eventId, bool expectedHasId, int expectedTypeFieldIndex)
-            {
-                Mock.Get(isOsPlatform).Setup(_ => _.Invoke(OSPlatform.Linux)).Returns(true);
-
-                var eventSource = new TestEventSource();
-
-                var eventDescriptorsField = typeof(ServiceFabricEventSource).GetField("eventDescriptors", BindingFlags.NonPublic | BindingFlags.Instance);
-                var eventDescriptors = (ReadOnlyDictionary<int, TraceEvent>)eventDescriptorsField.GetValue(eventSource);
-
-                Assert.True(eventDescriptors.ContainsKey(eventId));
-                var traceEvent = eventDescriptors[eventId];
-                Assert.Equal(expectedHasId, traceEvent.hasId);
-                Assert.Equal(expectedTypeFieldIndex, traceEvent.typeFieldIndex);
-            }
-
-            [Fact]
-            public void VariantWriteViaNative_OnNonLinux_ThrowsPlatformNotSupportedException()
-            {
-                Mock.Get(isOsPlatform).Setup(_ => _.Invoke(OSPlatform.Linux)).Returns(false);
-
-                var eventSource = new TestEventSource();
-                int eventId = 1;
-                int argCount = 3;
-                Variant v0 = "test";
-                Variant v1 = 42;
-                Variant v2 = true;
-
-                Assert.Throws<PlatformNotSupportedException>(() => eventSource.VariantWriteViaNative(eventId, argCount, v0, v1, v2));
-            }
         }
-
-#else
-        public sealed class NetFramework : ServiceFabricEventSourceTest
-        {
-            [Fact]
-            public void VariantWriteViaNative_NotAvailableOnNetFramework()
-            {
-                var eventSourceType = typeof(ServiceFabricEventSource);
-
-                var method = eventSourceType.GetMethod("VariantWriteViaNative", BindingFlags.Public | BindingFlags.Instance);
-
-                Assert.Null(method);
-            }
-        } 
-#endif
     }
 }
